--- sttapp/tasks/__init__.py
@@ -1,1 +1,0 @@
-from .mail_tasks import *
