--- sttapp/tasks/tasks.py
@@ -1,6 +1,0 @@
-from .celery import celery
-
-
-@celery.task()
-def add(a, b):
-    return a + b
